changed hasArrived
